Add debugging support to monitor what orca is thinking.
Change script management to be more like the a11y cache - simplifies
the orca code that works with scripts.

Provide support for the default script to be used in the event that
the application for an object cannot be found.

Provide support to enable/disable the loading and use of custom scripts
(allows one to verify the behavior of the default script).