in getDisplayedText, handle case where the text is a single
* src/orca/util.py: in getDisplayedText, handle case where the
text is a single EMBEDDED_OBJECT_CHARACTER.  In these cases,
we'll defer to getDisplayedText on the child for the text.  The
main driver for this is the Gecko toolkit, and this is a fix
to handle the majority of the problems I've been seeing with
labels in Gecko.  At some point, we may need to provide a more
sophisticated algorithm to handle strings containing one or
more EMBEDDED_OBJECT_CHARACTERs, but we'll cross that bridge
when we get to it.

svn path=/trunk/; revision=1902