hackndle (hack and handle) the case where things in the speech context are
* src/orca/util.py, src/orca/Gecko.py: hackndle (hack and handle)
the case where things in the speech context are objects that are
not our special objects that consist solely of a single
EMBEDDED_OBJECT_CHARACTER.  I'm going to need to bite the bullet
at some point and handle this whole EMBEDDED_OBJECT_CHARACTER
business.  It isn't going to be pretty.

svn path=/trunk/; revision=1916