Add setting to disable the pyatspi GIL idle handler
This change is made in response to
http://bugzilla.gnome.org/show_bug.cgi?id=576954 where is appears
as though the GIL idle handler in pyatspi is causing consumers of
pyatspi (e.g., Orca and accerciser) to start consuming 100% of a
CPU when the machine is idle.  We're pretty sure the GIL handler
is not necessary, so this change gets rid of it for the 2.27.1
development build of Orca.  If we see that users notice strange
lock ups, however, we may need to change things back.