Fix for bgo#581372 - Move cursor routing and six dot key handling from braille.py to script
With the cleanup from bgo#581532 to remove the custom brl module and
move directly to the brlapi module provided by BrlTTY/BrlAPI, we are
now able to handle BrlAPI commands much better.  This patch 'uncovers'
the handling the cursor routing keys and the six dot keys; they are
no longer swallowed/handled by the braille.py module alone.  Instead,
they go to the script like any other event.  By default, the script
just turns around and calls braille.py methods, but it can also feel
free to override them.  For example, it might look at keyboard
modifiers along with a cursor routing key to see if it wants to
select text or not.

As a note, the BrlAPI events come to us as a dictionary containing a
bunch of information about the event.  For example, the cursor routing
command contains information about which routing key was pressed.  The
six dots command, which is used to turn contracted braille on or off,
contains information about whether the user wants to turn contracted
braille on (dots 2-3-5 on my Baum display) or off (dots 2-3-6 on my
Baum display).  Right now, expressing interest in braille events is
still done by the command (e.g., brlapi.KEY_CMD_HOME,
brlapi.KEY_CMD_ROUTE, brlapi.KEY_CMD_SIXDOTS) and it is up to the
event handler to determine how to handle the arguments.