Fix for bgo#354471 - Text selection from braille input device
This is the first step of implementing this feature.  Here's the behavior:

KEY_CMD_CUTBEGIN (Dot 1 + cursor routing key on my display) - this will specify
the start of a selection.  Orca will merely move the caret to the given spot
and will clear any existing selection.

KEY_CMD_CUTLINE (Dot 4 + cursor routing key on my display) - this will specify
the end of a selection and the selected text is automatically copied to the
system clipboard.  If a selection doesn't exist, Orca creates a new one where
the other endpoint of the selection is where the caret is.  If a selection
exists and the selection point is outside the existing selection, Orca extends
the existing one.  If a selection exists and the selection point is inside the
existing selection, Orca trims the selection from the right (i.e., the selected
text that's after the selection point becomes unselected).

Known issues that need to be resolved:

1) This only works in text areas.  It doesn't work across things such as
   paragraphs in OpenOffice.

2) There's some strangeness with speech feedback: it sometimes says
   "unselected" when the text is selected.  This should be fixable, but
   there also probably shouldn't be any speech feedback when doing this
   from the braille display.