Work on bgo bug #574720
Work on bug #574720 - Table Navigation Keys for OpenOffice Writer.
This adds in support for Alt + the cursoring keys for navigation
in OOo tables. Note that you will need to first enable structural
by pressing Orca + Z. We plan to add in the customizations/settings
available in the Gecko preferences, as well as for the announcement
of dynamic row and column headers when navigating.