Refactor speech generators to be table driven
First phase of working on bug #Bug 570658 – Refactor the speech
and braille generators.  This phase touches the speech
generators and makes things very much table driven (see the
formatting.py modules).  Aside from touching many modules, the
impact on the user should *hopefully* be unnoticeable.  That is,
the new tables are set up to provide the same speech output we
were getting before the refactor.  The next phases will be
migrating where am I support to the speech generator and then
working on braille.