Part two of bug/RFE #574720
This adds settings for the table structural navigation. Note that
the setting to speak headings has been added, but it is not yet
implemented. OOo Writer doesn't expose headings as headings to us
currently. As a result, the only headings we will be able to speak
are those dynamically set by the user. Implementing that is part
three of this bug/RFE. :-)