Fix for bug 585871.
This fix moves some of the handling of embedded object characters
from the Gecko script to default.py. This change helps with solving
the immediate problem of bug 585871 (WhereAmI info in FF doesn't
include link text). It presumably will also come in handy as we begin
to deal with the presence of embedded object characters in other apps,
such as OOo Writer documents (e.g. forms) and potentially WebKit.