Fix for bgo#585897 - Pronunciation substitution can miss words due to word delimiters
This uses a regular expression to find word and non-word segments. It also
removes the fallback to chnames.py to get pronunciation for characters.
That is handled by the punctuation_settings dictionary, which is used
lower in the stack (e.g., in gnomespeechfactory.py).  As a result, the
SPEECH OUTPUT lines for those don't include the spoken-word form of
the characters and all the tests had to be changed back again.