Fix for bgo#578832 - Flash locking modifier key state in braille
This builds upon bgo#576848 to flash the state of the caps lock modifer
key.  It's currently bound to the settings.enableLockingKeys setting,
which is currently set via the key echo page.  If users want to
separate the presentation between speech and braille (e.g., don't
speak it, but braille it), we can make that change in the 2.29.x cycle.
Note also that as part of bgo#576848, orca.settings.brailleFlashTime
can be used to set the time (in milliseconds) that flashed messages
should appear before reverting back.  A negative number means to
flash indefinitely.  Finally, pressing a cursor routing key on the
braille display or a key on the keyboard will also revert the display
back to its pre-flashed state.