Fix for bgo#582491 - Strange message table behavior with Evolution
Note that this only handles the most extreme problem found - incoming
mail when Evo is not focused causing Orca to no longer provide access
to the app with focus. The within-Evo issues are due to bgo#483452
which was filed against Evolution in October 2007 and which has not
even been triaged by the Evolution maintainers. :-(