Fix for bgo#574344 - Support for different progress bar "verbosity levels" should be implemented
This adds a new setting, progressBarVerbosity, which specifies when
progres bar updates should be spoken, assuming they are enabled.
ALL means that any progress bar update will be spoken. APPLICATION
means that progress bar updates from the active application will be
spoken, even if they are not in the active window. WINDOW means that
the progress bar must be contained in the active window in order to
be spoken. The default verbosity level is APPLICATION, which was Orca's
behavior prior to this fix.

This also adds an option to the Orca Preferences dialog so that users
can choose a verbosity level without having to hand-edit their config
files.