Fix for bgo#523693 - Positioning of the cursor when editing text
This patch fixes a 'jumping cursor' problem when deleting text in the
middle of a long text area and also adds a _realignViewport method to
braille.py.  The method works off new settings:

ALIGN_BRAILLE_BY_EDGE   = 0
ALIGN_BRAILLE_BY_MARGIN = 1
ALIGN_BRAILLE_BY_WORD   = 2
brailleAlignmentMargin  = 3
brailleMaximumJump      = 8
brailleAlignmentStyle   = ALIGN_BRAILLE_BY_EDGE

The default alignment style matches what we have today.

The ALIGN_BRAILLE_BY_MARGIN style uses brailleAlignmentMargin and is
effectively a "push" model - when you get to the edge of the display, the
viewport is pushed to keep the cursor cell at the margin (until you reach the
edge of the text).

The ALIGN_BRAILLE_BY_WORD style uses brailleAlignmentMargin in the same push
model above, but when it pushes the viewport, it pushes it so the edge lands on
a word boundary.  In the event we hit a really long word, the
brailleMaximumJump setting limits how far we jump.