Most of the implementation for bgo#607663 - Create a script for Gajim.
This should cause all of the chat functionality and features in Orca
to work as expected.

Announcing the typing status does not yet work because I don't see a way
to distinguish statuses (typing, paying attention). This issue is seen in
the majority of clients and bugs will likely need to be filed against each
of those clients.

Similarly, it would be nice to present better information about the items
in the buddy list. I'll work on this next.