fix bgo#618589 - when debugging in eclipse, sometimes orca reads the same line twice
There are two situations in which this happened:
1. When a breakpoint is in a different file.
2. when a breakpoint is in different method.

in the first case the line wass announced twice because the events
object:state-changed:focused  and
object:text-caret-moved were received by orca.

in the second case the line wass announced twice because two events
object:text-caret-moved were received by orca.

The patch compares the object to be announced with the objectthe
announced previously.
If they identical and they contain the same caret offset, the
object is discarded.