fix order of dependancies in configure.in
the configure script used to check for pyatspi after orbit bonobo and dbus, however pyatspi requires these to function,
so if they are not present the script would say pyatspi could not be found when the problem was a dependancy was missing.
By checking for these first we can give a better error message.
Note we should probably have the at-spi configure.in check for the appropriate python moduels too.