fix for bgo#620018 - By reformatting a text or comment in eclipse, Orca speaks all the file contents
In the function onTextInserted, the event is discarded if  event.detail1 is
equals to 0 and event.detail2 is equals to
event.source.queryText().characterCount. When the text is reformated or
when a comment block is inserted, eclipse reinsert all the text but set
detail1 to 0 and detail2 to the length of the new text.