[config] Moved config files to XDG_DATA_HOME
Fix for bgo#596108 - Changed old ~/.orca config directory to XDG_DATA_HOME/orca which is more Freedesktop.org compilant.
There is a check for the old ~/.orca in case someone has an old configuration. A migration method will be provided for that, but
until this feature is implemented, the check is needed.