[config] Migration from ~/.orca to XDG-DATA-HOME option
Added provisional command line option for migrate the user's personal
preferences files from ~/.orca to XDG-DATA-HOME which is more fdo compilant.
By default Orca will use the old ~/.orca unless the user pass the '-m' or
'--migrate-config' option. This let's the betatesters to be using the last
Orca's version with the old config directory, so it doesn't mess with their
stable version's.