Fix for bug #630328 - Orca provides no access to WebKitGtk content whatsoever. Nada. Zilch.
Note: We are just now starting work on WebKitGtk and Epiphany scripts
in Orca. There will be tons more work that needs to be done. You
should not assume that this commit means we're done or that we are
not aware of remaining issues. Patience is a virtue. Remember to floss.
Have a nice day.