Adjustment to the fix for bug #520595 - Orca is far too "chatty" for persons with learning disabilities
Based on feedback from several sources, it is desirable to have feedback
from commands (time, date, learn mode, etc.) spoken even though that text
is not displayed on screen. This change causes all messages which are
presented via speakMessage or presentMessage to be spoken using the
system voice. We can make further adjustments as we get additional input
from the LD community on which messages should be silenced.