Fix for bug#642212 - Delay setting autostart until console UI is finished
Delay setting the gconf key to automatically start Orca until the console
UI is ready to exit, to avoid starting a new Orca instance while the console
setup is still running.

Use the settings manager accessor to set firstStart.

Silence speech when prompting for the default keyboard, to be consistent
with other prompts.