Work around for bug #665522 - Flat Review broken in Gecko apps with AT-SPI2
Quote from the orca-list:

"There's something going wrong with AT-SPI caching children. Either Mozilla
is sometimes not sending children-changed events when objects are added or
removed, or libatspi isn't properly updating its cache. I can't say which
for sure--I haven't taken the time to track it down.

This patch to Orca should help in the meantime."