Fix for bug 709447 - Orca segfaults when run with the wayland backend.
Sometimes Wnck.Screen.get_default() causes Orca to segfault when
Orca is launched with the wayland backend. It appears to be a race
condition, but it happens often enough and is severe enough that
I'll comment the offending call out for now so work related to
Wayland accessibility is not held up further.