Handle a missing event from a mystery object in the Evolution GUI
This is some mystery child of the 'Messages' panel which fails to show
up in the hierarchy or emit object:state-changed:focused events. But it
seems to be the only object that claims focus (via the deprecated focus:
event) when the 'Messages' tree table is navigated to. This really needs
to be fixed in Evo. But we're finally getting accessibility back, so who
am I to quibble? ;)