Validate reportedly selected children - needed for Evolution (at least)
When Evo's Mail Folders Tree gets focus, it reports having selected
children via the accessible selection interface. We were trusting that
those children were not dead, defunct, or otherwise bogus. As a result,
we were presenting nothing. This needs to be fixed in Evo, but doing a
sanity check will at least cause Orca to present the tree table.