Add optional start and end offset arguments for getTextLineAtCaret()
In the case of web content, we've already figured out and provided to
the generators the start offset, end offset, and string to be presented.
But then braille.py turns around and asks the script to (re)calculate it
and (re)provide it. This is 1000 different kinds of silly, and ultimately
getTextLineAtCaret() needs to go away. In the meantime, since multiple
scripts rely upon it (and for more than just braille), we might as well
make it possible for scripts to turn around and immediately hand back what
the braille code is asking for.