Don't set the caret position in Gecko zombie replicants
Gecko seems to like to kill certain accessible objects on us upon focus,
putting a new one in its place. This would completely mess up Orca's caret
navigation, so we added checks for these zombies and code to try to find
their replicants. That way, when the proverbial rug was snatched out from
under our feet, we could reposition ourselves on the replacement object,
and everything seemed fine. Now, however, the act of setting the caret in
the replicant seems to be sufficient to cause Gecko to kill that replicant.
This is why we can't nice things.

The change in this commit is to quietly update our locus of focus without
touching the zombie and its subsequent replicants in the hopes they'll
leave us alone so that we can actually navigate within the content.