Update our landmark role list to bring it into alignment with the ARIA spec
Our list of landmarks pre-dates the ARIA 1.0 REC by several years and
includes roles which are not landmarks and fails to include ARIA's form
role. We need to follow the spec. Note that the region role, which is not
a landmark role in ARIA 1.0, is a landmark role in ARIA 1.1. Hence its
continued inclusion in our list.