Don't look for zombie replicants in ROLE_EMBEDDED objects
If a new ROLE_EMBEDDED object shows up, odds are that the zombie locus
of focus is properly dead (as opposed to killed for sport by the app).
Plus, if that object has a giant and/or broken table, we can hang waiting
for pyatspi to get back to us.