Move some technical-content symbols to mathsymbols.py and remove duplicates
Because there are far more symbols than we can ask translators to translate,
we have fallback code so that we at least speak something for the symbols
which we think are unlikely to be encountered, and likely to not be presented
by espeak. This range-based code was stomping on a few symbols which were
actually localized. Moving them to mathsymbols.py will stop the stomping.
Also removed some duplicates.