Don't use pyatspi to cache children
Doing so seems to make us more dependent on app developers to create
accessible objects in a particular order. If they fail to do so, we
get an index in parent of -1 which makes the accessible look like a
candidate Zombie. Not caching children appears to work around this.
Whether or not it has side effects we don't like remains to be seen.
Unstable branch is unstable.