Comment out the name check for defunct accessibles
Simply asking for the name should not break anything and should make it
possible for us to quickly identify defunct accessibles whose events we
want to ignore and ultimately will ignore. But the mere act of asking
the event source for its name causes Orca to stop presenting Eclipse
(and possibly other) applications. :(