Present full contents of math tables that are part of a MathML expression
Unlike text-content tables, math tables can appear in the middle of an
expression. Not presenting the full table as part of the expression makes
the resulting expression at best confusing, and arguably wrong. The other
AT solutions are presenting the full math table so we will too for now. We
can always tweak later.