Don't insist upon label visibility for status bar notifications
Status bar notification support is our way of providing application
developers a sane, predictable means to have Orca notify users of
status changes and other messages. But having a visible status bar
and labels may be seen as an annoyance for sighted users. So we'll
not insist on label visibility in this particular case.