Update the soffice script to work better with setLocusOfFocus and onCaretMoved
The changes to setLocusOfFocus and onCaretMoved made Orca far more tolerant
regarding what to treat as a valid updates. As a result, Orca is now less
likely to fail to present something. The soffice script, however, was forcing
updates to ensure something would be spoken. That needed to be removed.