Add customization to check for AT-SPI2 registry crash and shutdown if detected
It appears that Orca may hang when the AT-SPI2 registry crashes. What is
triggering the registry crash needs to be figured out. In the meantime,
it's better to shutdown (and announce that we are doing so) than to hang.

Note that because this is a hack, and an in-progress one at that, polling
to verify the registry is still alive is disabled by default. It can be
enabled at the moment by adding the following lines to orca_customizations.py:

import orca.settings
orca.settings.checkForRegistryCrash = True

Note that this customization may change or go away at any time.