Revert "Add customization to check for AT-SPI2 registry crash and shutdown if detected"
The hack is no longer needed. We were trying to catch the SIGSEGV.
We shouldn't do that. If we don't try to catch it, when the registry
crashes or an invalid operation occurs in AT-SPI2, Orca will crash
properly. Of course, crashing is bad. But Orca has very little control
over AT-SPI2 bugs. And crashing, along with providing useful information,
beats hanging and providing no clues.

This reverts commit 2dba547bcae49f25649c185498d402776a464ea5.