Attempt to hack around SeaMonkey's broken events for HTML input popups
The popup claims focus when the user is inserting text in the input.
The input claims focus when the user is navigating within the popup.
And the popup is otherwise inaccessible, failing to emit focus or
selection change events. We cannot fix this, but we can try to make
it suck a bit less by ignoring popup focus claims when the user is
inserting text in the input.