Remove .desktop file
Because launching Orca does not cause any window to appear, Orca is
not seen as an "application." Furthermore, Orca can be started via
shortcut (Alt+Super+S) and listens to the "screen-reader-enabled" key
of the "org.gnome.desktop.a11y.applications" schema. Thus the .desktop
file was seen as no longer needed.

That said, just to be safe, we kept the .desktop file and used the
"NotShowIn" property to disable it from appearing in the environments
who didn't want it (e.g. GNOME and Unity). But that proved to be
insufficient because Orca was still showing up in distros' "App Store"
applications, leading to still more discussion about whether app stores
should hack around non-app applications or maintain separate packages
for Orca's .desktop file. All of this for a file which doesn't serve
much purpose for Orca's actual users because in order to launch Orca
from a desktop environment's Applications menu or search functionality,
that menu or search functionality needs to be accessible, which means
Orca needs to already be running.

If there are still distros which needs Orca's .desktop file, those
distros are encouraged to maintain it downstream.

https://bugzilla.gnome.org/show_bug.cgi?id=735761