Add explicit do-not-descend-for-caret check for ROLE_IMAGE
Images should not implement the accessible text interface so why bother
checking? Answer: If ARIA role="img" is used, it changes the role, but
not the interfaces implemented. Thus it is possible to have an image whose
child is a document. If the author tells us something is an image, we need
to trust the author.