Handle LO's recently-corrected exposure of text attributes for spellcheck
Note: The LO fix has only just been committed and is not yet available
in shipping versions of LO. In addition, in recent, earlier versions of
LO, a broken LO accessibility tree made it impossible for Orca to find
the spellcheck dialog. If your version of LO does not have fixes for both
of these issues, this change to Orca will not work.