Remove old column-header presentation suppression for Thunderbird
We were suppressing it due to chattiness when we had to put together
the full row from its cells. Accessible rows are now exposed to us
and we present those instead of the cells. Problem solved. And it
seems like this suppression interferes with presentation of valid
headers we wish to present.