Make sanity checking when getting scripts optional
Apparently fixing the bogus-multiple accessible-application instances
had the unintended side effect of making certain apps less responsive.
We need to try to figure out the magic combination to prevent double-
speaking without this side effect. NOTE: This particular change does not
change the current behavior or fix; it merely makes it possible to turn
checking off, which will be done on a trial-and-error basis in subsequent
commits.