braille: fix bogus dots appearing after text
endPos-startPos may not be as big as the braille output (which is the reason
for the whilewriteStruct.regionSize < _displaySize[0] loop actually).
attributeMask was getting grown by the loop, but then still cut to
endPos-startPos, thus leaving undefined values in the rest of the
braille output.

This changes the logic to the same as for computing substring: cut the
piece first, add spaces, and just put the result in the writeStruct.

https://bugzilla.gnome.org/show_bug.cgi?id=795015