braille: Accept offset being exactly at end of text
If the offset is exaclty at the end of some content, instead of returning
fallback values, we should rather return the content just before it, which
is notably needed when the caret is right after the current text.

We however do that only after having really failed to find an interval
containing the offset.