speech: add progress marking support
This introduces marks in the text sent to speech-dispatcher so that
speech modules which support indexing provide progress feedback. Scripts
can then use it from __sayAllProgressCallback to show the progression.

We have to do this in two phases: first put marks in the text, then let
various layers mangle it, and eventually use ssml to express the marks.