provide feedback on selecting all entries in a table
When a user presses ctrl-a to select all entries in a table (e.g. all files
in a Caja window), it would be useful to provide feedback, otherwise
nothing seems to have happened since for instance the focus hasn't
changed.

This adds a message in such case, in a way similar to doing the same for
a text entry.