Avoid double character representation expansion
`Script.pseakCharacter()` did the `chnames` expansion before passing it
to the speech module, which also expanded it tries and check if any
expansion happened, which is corrupted by the previous expansion.

This fixes potential double expansion, as well as allowing the speech
module to perform differently depending on whether the character has a
representation or not.