Add support for settings migration
This adds support for adding migration logic when loading user settings
from an older version.

If the user settings version doesn't match the running settings version
the migration methods kick in, with full power to transform options.

To implement a migration path for an option, implement the method
SettingsMigration._migrateOptName(self, userSettings, name)
which should transform userSettings as appropriate and return it.  It
allows any transformation, including renaming a setting.