Prevent app switcher speech from getting interrupted unexpectedly
This prevents the default window:deactivate handler from stopping the
app switcher speech.  It happens because as the switcher doesn't get
"activated" and the window:deactivate thinks it is interrupting its
own speech.  Work around that by removing the locus of focus so that
the window:deactivate handler doesn't think it owns the focus.