Present the full name for off-screen links
Some web publishers (e.g. Amazon) put off-screen links in their content
and keep them off-screen for screen reader users to find. Unfortunately,
when we ask for the text of these links by line, Gecko gives us a single
word at a time. In addition, when we try to work around this by using the
character extents, we get the same results. The safest thing to do for
these things is return the accessible name as the line content so users
don't have to press Down Arrow a zillion times to get past it.