Remove table row from focus-mode widget roles; grid logic should handle it
The focusModeRoles list contains roles which may be focus-mode or may not
be. The decision is based on whether or not the role is a text-block element.
A table row will never be a text block element. In addition, we have logic
which treats all grid descendants as focus-mode widgets, so that should
handle the most common use case for table row being a focus-mode widget.