Don't treat an element as a text block element if it has no text
The logic associated with text block elements is based on the fact that
there is static text to be had in these elements and thus the object
shouldn't' be treated as a whole. That fails for things like table cells
in a non-interactive table which have an author-provided name in lieu of
text.