Allow scripts to determine if we should really synthesize a click on a char
Some objects officially implement the accessible text interface, but don't
actually have meaningful text. In those cases, we don't want to try to
synthesize a click on a character; instead we should click on the object.
The scripts know what object really have text that should be handled on
a per-character basis so ask them if we have valid characters to click on.