Remove unused support for AtspiComponent's scrollToPoint
This support was added along with the rest of the component-based
scrolling API, but we're not using it. And at the present time, I
don't see that changing. We do have a need for, and are currently
using, the ability to scroll an object into view and to a particular
location (e.g. top edge). But I cannot see a reason why a screen
reader would need to scroll a particular object to a particular
point. Furthermore, with the new AtspiText's scrollSubstringTo,
we can ensure that a given substring is scrolled into view and even
scrolled to a particular location. We'll rely upon that instead of
AtspiComponent's scrollToPoint.