Just rely on the superscript's clearTextSelection() for web content
We were doing extra work to clear the text selection in web content
in anticipation of being able to have AT-SPI2-driven text selection
for web content. It turns out we cannot due to limitations in ATK
and assumptions of implementors. Therefore, don't bother doing the
extra work until we have a need to.