Chromium: Ignore "click-ancestor" action when identifying ignorable elements
Click-ancestor is meant to bubble up to an ancestor which is actually
clickable. But attempting to perform this action doesn't reliably work;
and the clickable ancestor is what we want to click on anyway. Treating
this as a valid action is causing us to include otherwise ignorable
elements such as sections with no semantic meaning.