descendantAtPoint: first only look for descendant with "showing" state
This is needed for e.g. evince which has several pages drawn at the same
place, but only one has the "showing" state.

Normally all widgets that can be pointed at should have "showing" state,
but some might be missing it and this change would introduce a regression
in such a case, so this change keeps the fallback of looking at descendants
which don't have "showing" state too.

Fixes #54