Try to filter out more Thunderbird event spam
If focus is on a folder and lots of mail comes in, we get two accessible
name change events (one for the row, one for the cell) for each and every
new message which came in. One. At. A. Time. These events are technically
not wrong, but they are really unfortunate.