Remove hackaround related to pages whose authoring messes up bounding boxes
The custom descendant at point was designed to work around authoring
which results in parent elements not containing their children as far
as the bounding boxes are concerned. But we keep finding more and more
edge cases. And the tree-dive solution was not working. Therefore, we
need to get this fixed in the user agent.