Re-create the Xmodmap on keyboard changes
When a new keyboard is plugged in, the Xmodmap gets reset.
This stomps on on our changes to the Xmodmap (related to Orca modifier keys).
This change introduces event handlers that listen to 'device-added' and 'device-removed'
events. If the device being added or removed is a keyboard, we re-create our key mappings.