Chromium: Update support for list item marker change
Chromium used to expose list item markers to us as static text leaf
nodes; now they appear in a container. This makes it possible to
simplify our support, but we also need to eliminate presentation of
the new panel role we're seeing.

Also improve logic regarding what is a list item marker so we do
not double-present text in an inline list.