Don't call saveFocusedObjectInfo() in setCaretPosition()
When we update our position via caret navigation and then present the
new location, we lose state information that we need, such as the cell
coordinates we used to be in. This can cause us to think the row hasn't
changed when, in fact, it has.

We're already saving state in locusOfFocusChanged(). That should be the
only place we do it.