Revert "Don't call saveFocusedObjectInfo() in setCaretPosition()"
This reverts commit 716d2b9e099e2e2cea2a94f1f8ab5f26ce277fec.

When we set the caret position and update the locusOfFocus without
notifying the script, state isn't getting saved. We should consider
refining setLocusOfFocus to update all state even when the script
shouldn't be "notified" (which really means "don't present this change
to the user"). Doing that cleanly is a bigger change. So reverting for
now.