If the line contents start with a newline char, remove the newline char
Some user agents report the newline character shares the same bounding
box as the contents which follow on the same line. We don't want to
present that newline character. Also setting the caret at that point
typically places it at the end of the previous line rather than at the
beginning of the line we're about to present.