shiftlock: Never drop shift from keybindings
When setting a keybinding, orca tries to drop the underlying modifier of the
Orca modifier from the shortcut, if any, since that modifier is not supposed
to show up in the keyboard event, thanks to the keyboard modification.

We however can not do this for the shift modifier (in the case xkb is
configured with caps:shiftlock instead of caps:capslock), because then
if the caps:capslock case, shortcuts which use orca+shift+key would get
their shift modifier dropped.

The underlying modifier is not supposed to show up here anyway (since we
tinker with the xkb setup), so better not even try to fix them, it
brings more harm than good.

Fixes #87